separate registration and verification forms into sub components
